Move race condition utilities to test tool infrastructure
This is a preparatory refactoring PR so we can use the same utilities
for the existing race condition tests and for a test suite that tests
race conditions in combinations with exceptions.

changelog_begin
changelog_end
